+++ src/renderer/components/SearchInput.tsx
@@ -5,16 +5,20 @@ import { useTranslationFunction } from '../contexts'
 export function ClearButton(props: {
   onChange: (event: { target: { value: '' } }) => void
   value: string
+  extraCleanAction?: () => void
 }) {
-  const { onChange, value } = props
-  const onClear = () => onChange({ target: { value: '' } })
+  const { onChange, extraCleanAction, value } = props
+  const onClear = () => {
+    onChange({ target: { value: '' } })
+    extraCleanAction?.()
+  }
 
   return (
     <button
       aria-label='Clear'
       className={classNames(
         'bp4-dialog-close-button bp4-button bp4-minimal bp4-icon-large bp4-icon-cross clear-button',
-        { 'clear-button--hidden': value === '' }
+        { 'clear-button--hidden': value === '' && !extraCleanAction }
       )}
       onClick={onClear}
     />
@@ -29,8 +33,10 @@ export default function SearchInput(props: {
   className: string
   id: string
   inputRef?: React.ClassAttributes<HTMLInputElement>['ref']
+  /** if this is defined clear button is always shown, like with search in chat */
+  extraCleanAction?: () => void
 }) {
-  const { onChange, value, className, id } = props
+  const { onChange, value, className, id, extraCleanAction } = props
   const tx = useTranslationFunction()
   return (
     <>
@@ -44,7 +50,11 @@ export default function SearchInput(props: {
         ref={props.inputRef}
         spellCheck={false}
       />
-      <ClearButton value={value} onChange={onChange} />
+      <ClearButton
+        value={value}
+        onChange={onChange}
+        extraCleanAction={extraCleanAction}
+      />
     </>
   )
 }